+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## 0.6.1
 
-* Adding rebuildSchema to createCollection
-* Adding purge to removeCollection
+* Adding rebuild to createCollection
+* Adding drop to removeCollection
 
 ## 0.6.0